Add basic framework skeleton for frontend and backend
The backend is made with sinatra and can be developed and served by
using docker. For development the source directory can just be bind
mounted into the container.

The frontend is made with ember and compiled at docker build time. For
development it can be `ember server --proxy 127.0.0.1:3000`'d if you
bind the backend containers internal 4567 port to 3000.

Another option for develoment is to `ember build --watch` and then bind
mount the `assets/dist` directory into the containers /app/public
directory.